Damas v1.1
This is version 1.1, many classes and modules have been modified, restructured and made more efficient. The overall look of the game is the same as 1.0 but the graphic engine now runs much smoother and more reliable. The next step is the development of a solid AI module that will be unbeatable by a human.